# openapi/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Gestionale Negozio API
  description: API del gestionale negozio
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/stato-riparazione:
    get:
      tags:
      - stato-riparazione-controller
      operationId: getAllStatoRiparazione
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatoRiparazioneResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      tags:
      - stato-riparazione-controller
      operationId: updateAllStatoRiparazione
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatoRiparazioneRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatoRiparazioneResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      tags:
      - stato-riparazione-controller
      operationId: createStatoRiparazione
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStatoRiparazioneRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatoRiparazioneResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /api/stato-riparazione/{id}:
    get:
      tags:
      - stato-riparazione-controller
      operationId: getStatoRiparazioneById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatoRiparazioneResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      tags:
      - stato-riparazione-controller
      operationId: deleteStatoRiparazione
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
    ApiError:
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    StatoRiparazioneItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
          minLength: 1
        colore:
          type: string
          minLength: 1
        position:
          type: integer
          format: int32
      required:
      - colore
      - id
      - nome
      - position
    UpdateStatoRiparazioneRequest:
      type: object
      properties:
        stati:
          type: array
          items:
            $ref: "#/components/schemas/StatoRiparazioneItem"
      required:
      - stati
    StatoRiparazioneResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
          minLength: 1
        colore:
          type: string
          minLength: 1
        position:
          type: integer
          format: int32
      required:
      - colore
      - id
      - nome
      - position
...
    CreateStatoRiparazioneRequest:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        colore:
          type: string
          minLength: 1
        position:
          type: integer
          format: int32
      required:
      - colore
      - nome
      - position
...